# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart AI Prediksi Togel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /predict/latest:
    get:
      operationId: getLatestPrediction
      tags: [predict]
      summary: Get latest stored prediction
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
      responses:
        "200":
          description: Latest prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
        "404":
          description: No prediction found
  /predict:
    post:
      operationId: runPrediction
      tags: [predict]
      summary: Run AI prediction using 100 engines
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionInput"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
  /predict/accuracy:
    get:
      operationId: getPredictionAccuracy
      tags: [predict]
      summary: Get prediction accuracy statistics
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
      responses:
        "200":
          description: Accuracy statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionAccuracyStats"
...
  /learning/log:
    get:
      operationId: getLearningLog
      tags: [learning]
      summary: Get self-learning evaluation log
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Learning log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningLogResponse"
  /learning/performance:
    get:
      operationId: getEnginePerformance
      tags: [learning]
      summary: Get engine performance statistics
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Engine performance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnginePerformanceResponse"
...
components:
  schemas:
...
    AccuracyRecord:
      type: object
      required: [days, totalChecked, hit4d, hit3d, hit2d, hitBbfs6, hitColokBebas, winRate4d, winRate3d, winRate2d, winRateBbfs6, winRateColokBebas]
      properties:
        days:
          type: integer
        totalChecked:
          type: integer
        hit4d:
          type: integer
        hit3d:
          type: integer
        hit2d:
          type: integer
        hitBbfs6:
          type: integer
        hitColokBebas:
          type: integer
        winRate4d:
          type: number
        winRate3d:
          type: number
        winRate2d:
          type: number
        winRateBbfs6:
          type: number
        winRateColokBebas:
          type: number
    RecentScore:
      type: object
      required: [date, actual, hit4d, hit3d, hit2d, hitBbfs6, hitColokBebas]
      properties:
        date:
          type: string
        actual:
          type: string
        hit4d:
          type: boolean
        hit3d:
          type: boolean
        hit2d:
          type: boolean
        hitBbfs6:
          type: boolean
        hitColokBebas:
          type: boolean
    PredictionAccuracyStats:
      type: object
      required: [pasaran, totalPredictions, totalScored, last7, last30, last90, recentScores]
      properties:
        pasaran:
          type: string
        totalPredictions:
          type: integer
        totalScored:
          type: integer
        last7:
          $ref: "#/components/schemas/AccuracyRecord"
        last30:
          $ref: "#/components/schemas/AccuracyRecord"
        last90:
          $ref: "#/components/schemas/AccuracyRecord"
        recentScores:
          type: array
          items:
            $ref: "#/components/schemas/RecentScore"
...
    EngineItem:
      type: object
      required: [engineId, engineName, category, digits, confidence, weight]
      properties:
        engineId:
          type: integer
        engineName:
          type: string
        category:
          type: string
        digits:
          type: object
          properties:
            as:
              type: array
              items:
                type: string
            kop:
              type: array
              items:
                type: string
            kepala:
              type: array
              items:
                type: string
            ekor:
              type: array
              items:
                type: string
        confidence:
          type: number
        weight:
          type: number
    EngineContribution:
      type: object
      required: [category, contribution]
      properties:
        category:
          type: string
        contribution:
          type: number
    TopEngineItem:
      type: object
      required: [name, category, weight]
      properties:
        name:
          type: string
        category:
          type: string
        weight:
          type: number
    PredictionResult:
      type: object
      required: [pasaran, generatedAt, consensus4d, consensus3d, consensus2d, colokBebas, bbfs5, bbfs6, overallConfidence, engines]
      properties:
        pasaran:
          type: string
        generatedAt:
          type: string
        totalDrawsUsed:
          type: integer
        engines:
          type: array
          items:
            $ref: "#/components/schemas/EngineItem"
        consensus4d:
          type: array
          items:
            type: string
        consensus3d:
          type: array
          items:
            type: string
        consensus2d:
          type: array
          items:
            type: string
        colokBebas:
          type: array
          items:
            type: string
        bbfs5:
          type: array
          items:
            type: string
        bbfs6:
          type: array
          items:
            type: string
        overallConfidence:
          type: number
        engineSummary:
          type: object
          additionalProperties:
            type: number
        explanations:
          type: array
          items:
            type: string
        engineContributions:
          type: array
          items:
            $ref: "#/components/schemas/EngineContribution"
        topEngines:
          type: array
          items:
            $ref: "#/components/schemas/TopEngineItem"
    PredictionInput:
      type: object
      required: [pasaran]
      properties:
        pasaran:
          type: string
        limit:
          type: integer
...
    LearningLogItem:
      type: object
      required: [id, actualResult, hit4d, hit3d, hit2d, hitBbfs6, hitColokBebas, createdAt]
      properties:
        id:
          type: integer
        actualResult:
          type: string
        hit4d:
          type: boolean
        hit3d:
          type: boolean
        hit2d:
          type: boolean
        hitBbfs6:
          type: boolean
        hitColokBebas:
          type: boolean
        asHit:
          type: boolean
        kopHit:
          type: boolean
        kepalaHit:
          type: boolean
        ekorHit:
          type: boolean
        predicted4dTop5:
          type: array
          items:
            type: string
        bestCategory:
          type: string
        worstCategory:
          type: string
        bestCategoryScore:
          type: number
        worstCategoryScore:
          type: number
        categoryScores:
          type: object
          additionalProperties:
            type: number
        weightAdjustments:
          type: object
          additionalProperties:
            type: number
        learningNote:
          type: string
        createdAt:
          type: string
    LearningLogResponse:
      type: object
      required: [pasaran, log]
      properties:
        pasaran:
          type: string
        log:
          type: array
          items:
            $ref: "#/components/schemas/LearningLogItem"
    EnginePerformanceItem:
      type: object
      required: [key, category, avgScore, multiplier, trend, evaluationCount]
      properties:
        key:
          type: string
        category:
          type: string
        avgScore:
          type: number
        multiplier:
          type: number
        trend:
          type: string
        evaluationCount:
          type: integer
    WeightTableItem:
      type: object
      required: [key, category, multiplier, sampleSize]
      properties:
        key:
          type: string
        category:
          type: string
        multiplier:
          type: number
        sampleSize:
          type: integer
    EnginePerformanceResponse:
      type: object
      required: [pasaran, performance, weights]
      properties:
        pasaran:
          type: string
        performance:
          type: array
          items:
            $ref: "#/components/schemas/EnginePerformanceItem"
        weights:
          type: array
          items:
            $ref: "#/components/schemas/WeightTableItem"
...